c-automatisation-test/requirements.txt
numpy==1.23.1
pandas==1.5.0
pytest
unittest2
mysql.connector
fastapi
